Changed minimal tolerance and steps, changed throttle to double to improve precision when calculating high power outputs.

+++ Source_VEHelixien/VEHelixienIntegration.cs
@@ -24,8 +24,8 @@ namespace PowerController
 			CompPowerController Controller = __instance.parent.GetComp<CompPowerController>();
 			if (Controller != null)
 			{
-				__result *= Controller.Throttle;
+				__result *= (float)Controller.Throttle;
 			}
 		}
-	}	
+	}
 }
